==> Capstone/data cleasing/01. Output under score columns.ipynb <==
##### upload csv file

# import numpy and pandas
import numpy as np
import pandas as pd

# used for dates
import datetime
from datetime import datetime, date

# Set some pandas options controlling output format
pd.set_option('display.notebook_repr_html', False)
pd.set_option('display.max_columns', 8)
pd.set_option('display.max_rows', 100)
pd.set_option('display.width', 90)

# bring in matplotlib for graphics
import matplotlib.pyplot as plt
%matplotlib inline

# read in msft.csv into a DataFrame
msft = pd.read_csv("data\member_sample.csv")
msft[:15]

df = msft.copy()

df.shape

clist=[]
len = df.columns.shape[0]
for i in range(1,len):
    temp = df.columns[i]
    clist.append(temp.replace(" ","_"))

#print("column list:", clist)    

# remove unwanted column 0
df1 = df.copy()

col = df1.columns[0]
del df1[col]

# rename column names:
df1.columns = clist

df1.columns

# x = df1['Occupation_Group'].dropna()
# x.head()

# write the output to data sub-directory
root = 'data'
df1.to_csv( root + '/' +'1.member_column_cleanup.csv',index=False)
