==> Capstone/data cleasing/02. Output Include Member Tenure Tier.ipynb <==
##### upload csv file

# import numpy and pandas
import numpy as np
import pandas as pd
from sklearn import preprocessing

# used for dates
import datetime
from datetime import datetime, date

# Set some pandas options controlling output format
pd.set_option('display.notebook_repr_html', False)
pd.set_option('display.max_columns', 8)
pd.set_option('display.max_rows', 100)
pd.set_option('display.width', 90)

# bring in matplotlib for graphics
import matplotlib.pyplot as plt
%matplotlib inline

# read output #1 (repalce space with underscore in all columns)
root = 'data'
msft = pd.read_csv( root + '/' +'1.member_column_cleanup.csv')
#msft[:15]

df = msft.copy()

df.shape

# split na from memeber tenure years before binning data
# add df_na back at the end of binning data process
df_na = df[df['Member_Tenure_Years'].isnull()].copy()
df_temp = df[df['Member_Tenure_Years'].notnull()].copy()

# print(df_na.shape)
# df_na.transpose()
# # forget this row since it has not much data in it
# # no need to add it back

df_temp.columns

x = df_temp['Occupation_Group'].dropna()
x.head()

li = df_temp['Member_Tenure_Years'].unique().tolist()
#sorted(li)

max_value = df_temp['Member_Tenure_Years'].max()
print('max value:', int(max_value))

# Deciles as integer from 1 to 10
df_temp['Member_Tenure_Decile'] = pd.qcut(df_temp['Member_Tenure_Years'],q=10,labels=False) + 1
df_temp['Member_Tenure_Decile'].value_counts().sort_index()

###### The iter() function takes two parameters:

object - object whose iterator has to be created (can be sets, tuples, etc.)

sentinel (optional) - special value that is used to represent the end of a sequence

# # convert list to dictionary
# lst = ['a', 1, 'b', 2, 'c', 3] 
# it=iter(lst)
# res_dct = dict(zip(it, it)) 
# print(res_dct)

# assign memeber tenure tier

category,bins = pd.qcut(df_temp['Member_Tenure_Years'],q=10,retbins=True,labels=False)

bins_lst=[]
for i in range(0,len(bins)):
    if i == 0:
        bins_lst = [1,'0-10']
    elif i+1 < len(bins):
        k = (i+1)
        bins_lst.append(k)
        v = str(int(bins[i])+1) + "-" + str(int(bins[i+1]))
        bins_lst.append(v)
        
#print(bins_lst)
it=iter(bins_lst)
bins_dct = dict(zip(it, it)) 
print(bins_dct)

# repalce category key with a more describe names by using map function
# add a new column to df_temp
df_temp['Member_Tenure_Tier'] = df_temp['Member_Tenure_Decile'].map(bins_dct)
#df_na['Member_Tenure_Tier'] = np.nan

#df_temp['Member_Tenure_Tier'].unique()[:15]
#df_na['Member_Tenure_Tier']

# spike_cols = [col for col in df_temp.columns if 'Decile' in col]
# print(spike_cols)
# spike_cols = [col for col in df_na.columns if 'Decile' in col]
# print(spike_cols)

# delete Member_Tenure_Decile column from df_temp before merge it back with df_na
# del df_temp['Member_Tenure_Decile']
# spike_cols = [col for col in df_temp.columns if 'Decile' in col]
# print(spike_cols)

# df1 = pd.concat([df_temp, df_na], ignore_index=True, sort=True)
# print("df_temp:", df_temp.shape)
# print("df_na:", df_na.shape)
# print("df1.shape", df1.shape)

df_temp.columns

x1 = df_temp['Occupation_Group'].dropna()
x1.head()

# write the output to data sub-directory
root = 'data'
df_temp.to_csv( root + '/' +'2.member_tenure_tier.csv',index=False)
