==> Capstone/data cleasing/03. Output Products Label Binarizer.ipynb <==
##### upload csv file

# import numpy and pandas
import numpy as np
import pandas as pd
from sklearn import preprocessing

# used for dates
import datetime
from datetime import datetime, date

# Set some pandas options controlling output format
pd.set_option('display.notebook_repr_html', False)
pd.set_option('display.max_columns', 8)
pd.set_option('display.max_rows', 100)
pd.set_option('display.width', 90)

# bring in matplotlib for graphics
import matplotlib.pyplot as plt
%matplotlib inline

# read output #2 (new column for memeber tenure tier)
root = 'data'
msft = pd.read_csv( root + '/' +'2.member_tenure_tier.csv')
#msft[:15]

df = msft.copy()

df.shape

x2 = df['Occupation_Group'].dropna()
x2.head()

spike_cols = [col for col in df.columns if 'Tier' in col ]
print(spike_cols, "\n")

spike_cols = [col for col in df.columns if 'Flag' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'FSV' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'TRV' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'Children' in col ]
print(spike_cols, "\n")

df[df['FSV_CMSI_Flag'].isnull()].shape[0]

# verify the value in the columns:
a = np.array(['FSV_CMSI_Flag', 'FSV_Credit_Card_Flag', 'FSV_Deposit_Program_Flag', 'FSV_Home_Equity_Flag', 'FSV_ID_Theft_Flag', 'FSV_Mortgage_Flag', 'TRV_Globalware_Flag', 'Children'])
for i in range(0,len(a)):
    temp = a[i]
    nan = df[df[a[i]].isnull()].shape[0]
    print("current state " + a[i] + " values:", df[temp].unique(), " and nan = ", nan, "\n")
    

from sklearn import preprocessing

# Convert binary fields to numeric boolean fields
lb = preprocessing.LabelBinarizer()
df_dummy = df.copy()
a = np.array(['FSV_CMSI_Flag', 'FSV_Credit_Card_Flag', 'FSV_Deposit_Program_Flag', 'FSV_Home_Equity_Flag', 'FSV_ID_Theft_Flag', 'FSV_Mortgage_Flag', 'TRV_Globalware_Flag'])
for i in range(0,len(a)):
    temp = a[i]
    df_dummy[temp] = lb.fit_transform(df_dummy[temp])

#df_dummy['FSV_CMSI_Flag']

# verify the value in the columns:
a = np.array(['FSV_CMSI_Flag', 'FSV_Credit_Card_Flag', 'FSV_Deposit_Program_Flag', 'FSV_Home_Equity_Flag', 'FSV_ID_Theft_Flag', 'FSV_Mortgage_Flag', 'TRV_Globalware_Flag'])
for i in range(0,len(a)):
    temp = a[i]
    print(df_dummy[temp].unique())

df_dummy.shape

x2 = df_dummy['Occupation_Group'].dropna()
x2.head()

# write the output to data sub-directory
root = 'data'
df_dummy.to_csv( root + '/' +'3.product_label_binarizer.csv',index=False)
