==> Capstone/data cleasing/04. Output Total Products.ipynb <==
##### upload csv file

# import numpy and pandas
import numpy as np
import pandas as pd
from sklearn import preprocessing

# used for dates
import datetime
from datetime import datetime, date

# Set some pandas options controlling output format
pd.set_option('display.notebook_repr_html', False)
pd.set_option('display.max_columns', 8)
pd.set_option('display.max_rows', 100)
pd.set_option('display.width', 90)

# bring in matplotlib for graphics
import matplotlib.pyplot as plt
%matplotlib inline

# read output #2 (new column for memeber tenure tier)
root = 'data'
msft = pd.read_csv(root + '/' +'3.product_label_binarizer.csv')
#msft[:15]

df = msft.copy()

df.shape

spike_cols = [col for col in df.columns if 'Tier' in col ]
print(spike_cols, "\n")

spike_cols = [col for col in df.columns if 'Flag' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'FSV' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'TRV' in col ]
print(spike_cols, "\n")

a =df[['Individual_Key','FSV_CMSI_Flag', 'FSV_Credit_Card_Flag', 'FSV_Deposit_Program_Flag', 
 'FSV_Home_Equity_Flag', 'FSV_ID_Theft_Flag', 'FSV_Mortgage_Flag','TRV_Globalware_Flag']]
df_temp = a.drop_duplicates(subset=None, keep='first', inplace=False)
print(df.shape)
print(a.shape)
print(df_temp.shape)
df_temp[:10]
#df['Product_Count']

df_sum = df_temp.set_index('Individual_Key')

#Total sum per row: 
df_sum['Product_Count'] = df_sum.sum(axis=1)
df_sum.loc[52211550.0]

df_temp = df_sum.reset_index().copy()
df_temp.columns
df_total = df_temp[['Individual_Key', 'Product_Count']].copy()

# find the column index
df.columns.get_loc('TRV_Globalware_Flag')

# adding the product count into the orginal dataset
df_result = pd.merge(df, df_total, on='Individual_Key')
print(df_result.shape)
print(df_result.columns)

# add new column to the 1st position
df_result.insert(15, 'Total_Products', df_result['Product_Count'])
df_result['Total_Products']
del df_result['Product_Count']

# a client has max 4 products
print(sorted(df_result['Total_Products'].unique()))

print("# of Right_Individual_Key not same as Right_Individual_Key: ", \
      df_result[df_result['Right_Individual_Key'] != df_result['Individual_Key']].shape[0])


# delete the column
del df_result['Right_Individual_Key']

spike_cols = [col for col in df.columns if 'Member' in col ]
print(spike_cols, "\n")

print("unique member flag:", df_result['Member_Flag'].unique())
print("has nan?:", df_result[df_result['Member_Flag'].isnull()].shape[0])

# all the clients are AAA member - omit it
del df_result['Member_Flag']

df_result.columns

z1 = df_result['Occupation_Group'].dropna()
z1.head()

# # write the output to data sub-directory
root = 'data'
df_result.to_csv( root + '/' +'4.total_products_cleanup.csv',index=False)
