==> Capstone/data cleasing/05. Output Calculated Age.ipynb <==
##### upload csv file

# import numpy and pandas
import numpy as np
import pandas as pd
from sklearn import preprocessing

# used for dates
import datetime
from datetime import datetime, date

# Set some pandas options controlling output format
pd.set_option('display.notebook_repr_html', False)
pd.set_option('display.max_columns', 8)
pd.set_option('display.max_rows', 100)
pd.set_option('display.width', 90)

# bring in matplotlib for graphics
import matplotlib.pyplot as plt
%matplotlib inline

# read output #4 (file has total products)
root = 'data'
msft = pd.read_csv(root + '/' +'4.total_products_cleanup.csv')
#msft[:15]


df = msft.copy()

df.shape

spike_cols = [col for col in df.columns if 'Tier' in col ]
print(spike_cols, "\n")

spike_cols = [col for col in df.columns if 'Key' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'FSV' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'TRV' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'Total' in col ]
print(spike_cols, "\n")
spike_cols = [col for col in df.columns if 'Status' in col ]
print(spike_cols, "\n")


spike_cols = [col for col in df.columns if 'Birth' in col ]
print(spike_cols, "\n")

df['Date_Of_Birth'][:5]

# remove time from date + time format
df['Date_Of_Birth'] = pd.to_datetime(df['Date_Of_Birth']).dt.date
df['Date_Of_Birth'][:5]

print(datetime.today().strftime('%Y-%m-%d'))

#del woo

woo = df[['Individual_Key', 'Date_Of_Birth']].copy()

# remove warning message 
pd.options.mode.chained_assignment = None


tdy = datetime.today().strftime('%Y-%m-%d')
#Add a New Column to the DataFrame
woo.columns
woo.loc[:,"Now"]  = tdy
woo[:5]
#woo['Now'] = pd.to_datetime(tdy)

woo['Age'] = pd.to_datetime(woo['Now']).dt.year - pd.to_datetime(woo['Date_Of_Birth']).dt.year

# format age as int
# repalce nan with 0 to avoid error message
woo['Age'] = woo['Age'].replace(np.NaN, -1)
woo['Age'] = woo['Age'].astype(int)
woo[:10]


del woo['Date_Of_Birth']
woo[:5]

del woo['Now']
woo[:5]

# add an new column: Age to the original dataset
# adding the product count into the orginal dataset
df_result = pd.merge(df, woo, on='Individual_Key')

# add 1 as Individual count in the dataset
df_result['Individual_Cnt'] = 1

print(df_result.shape)
print(df_result.columns)


df_result['Individual_Cnt'].count()

df_result.rename(columns = {'Age':'temp'}, inplace = True) 

# find the column index
df_result.columns.get_loc('Date_Of_Birth')

# add new column to the 1st position
df_result.insert(23, 'Age', df_result['temp'])
woo['Age']
del df_result['temp']
df_result[:2]

df_result.shape

df.shape

# find the column index
df_result.columns.get_loc('Age')

df_result.columns

w1 = df_result['Occupation_Group'].dropna()
w1.head()

# # write the output to data sub-directory
root = 'data'
df_result.to_csv( root + '/' +'5.member_age.csv',index=False)
